Move menu item components into seperate dir.
