+++ graylog2-web-interface/src/components/bootstrap/DropdownButton.test.tsx
@@ -18,7 +18,7 @@ import * as React from 'react';
 import { render, screen, waitFor } from 'wrappedTestingLibrary';
 import userEvent from '@testing-library/user-event';
 
-import MenuItem from 'components/bootstrap/MenuItem';
+import { MenuItem } from 'components/bootstrap';
 
 import DropdownButton from './DropdownButton';
