+++ graylog2-web-interface/src/components/bootstrap/index.ts
@@ -24,7 +24,7 @@ export { default as ButtonGroup } from './ButtonGroup';
 export { default as ButtonToolbar } from './ButtonToolbar';
 export { default as Checkbox } from './Checkbox';
 export { default as ControlLabel } from './ControlLabel';
-export { default as DeleteMenuItem } from './DeleteMenuItem';
+export { default as DeleteMenuItem } from './menuitem/DeleteMenuItem';
 export { default as DropdownButton } from './DropdownButton';
 export { default as FormControl } from './FormControl';
 export { default as FormGroup } from './FormGroup';
@@ -38,7 +38,7 @@ export { default as Label } from './Label';
 export { default as ListGroup } from './ListGroup';
 export { default as ListGroupItem } from './ListGroupItem';
 export { default as Menu } from './Menu';
-export { default as MenuItem } from './MenuItem';
+export { default as MenuItem } from './menuitem/MenuItem';
 export { default as Modal } from './Modal';
 export { default as Nav } from './Nav';
 export { default as NavDropdown } from './NavDropdown';
