+++ graylog2-web-interface/src/components/bootstrap/menuitem/DeleteMenuItem.tsx
