+++ graylog2-web-interface/src/components/event-notifications/event-notifications/EventNotifications.tsx
@@ -32,9 +32,8 @@ import {
   Icon,
   QueryHelper,
 } from 'components/common';
-import { Col, DropdownButton, MenuItem, Row, Button } from 'components/bootstrap';
+import { Col, DropdownButton, MenuItem, Row, Button, DeleteMenuItem } from 'components/bootstrap';
 import Routes from 'routing/Routes';
-import DeleteMenuItem from 'components/bootstrap/DeleteMenuItem';
 
 import styles from './EventNotifications.css';
