+++ graylog2-web-interface/src/components/streams/StreamsOverview/BulkActions/SearchStreamsAction.tsx
@@ -17,7 +17,7 @@
 
 import * as React from 'react';
 
-import MenuItem from 'components/bootstrap/MenuItem';
+import { MenuItem } from 'components/bootstrap';
 import { DEFAULT_TIMERANGE } from 'views/Constants';
 import SearchLink from 'views/logic/search/SearchLink';
 import { createElasticsearchQueryString } from 'views/logic/queries/Query';
